unify artifact actions
